+++ lm_council/__init__.py
@@ -1,0 +1,1 @@
+from .council import LanguageModelCouncil
