+++ lm_council/judging/preset_configs/default_pairwise.yaml
@@ -1,0 +1,46 @@
+type: pairwise_comparison
+exclude_self_grading: false
+cot_enabled: false
+temperature: 0.0
+config:
+  prompt_template: 'Two models provided a response to a user prompt.
+    
+
+    ### USER PROMPT START ###
+
+    {user_prompt}
+
+    ### USER PROMPT END ###
+
+
+    We want to decide which response is better. To assses quality, think about coherence and relevance.
+
+
+    ### RESPONSE A START ###
+
+    {response_1}
+
+    ### RESPONSE A END ###
+
+
+    ### RESPONSE B START ###
+
+    {response_2}
+
+    ### RESPONSE B END ###
+
+
+    You must output only one of the following choices as your final verdict with a
+    label:
+
+    {pairwise_comparison_labels}
+
+    '
+  granularity: 4
+  skip_equal_pairs: true
+  algorithm_type: fixed_reference_models
+  position_flipping: true
+  algorithm_config:
+    reference_models:
+    - google/gemini-2.5-flash-preview-05-20
+reps: 1
